+++ .github/bin/release-plan/npm-can-publish.mjs
@@ -1,39 +1,7 @@
 import { spawn } from 'child_process';
 import { platform } from 'process';
 
-export async function canPublish(packageName) {
-	const myName = await new Promise((resolve, reject) => {
-		const whoamiCmd = spawn(
-			'npm',
-			[
-				'whoami'
-			],
-			{
-				shell: platform === 'win32',
-				stdio: 'pipe'
-			}
-		);
-
-		let result = '';
-
-		whoamiCmd.stdout.on('data', (data) => {
-			result += data;
-		});
-
-		whoamiCmd.on('close', (code) => {
-			if (0 !== code) {
-				reject(new Error(`'npm whoami' exited with code ${code}`));
-				return;
-			}
-
-			resolve(result.trim());
-		});
-	});
-
-	if (!myName) {
-		return false;
-	}
-
+export async function canPublish(packageName, myName) {
 	return new Promise((resolve, reject) => {
 		const accessListCmd = spawn(
 			'npm',
